Infra: PCH expansion + zero warnings + /WX

+++ StratEngine/AIBasic.cpp
@@ -44,7 +44,7 @@ int AIBasic::Search(_In_ size_t ply, _In_ int alpha, _In_ int beta)
 	const GameInfo& info = GetLastBoardInfo(ply);
 
 	// Test for 50 moves rule
-	if (checkDraws(info, ply))
+	if (checkDraws(info, static_cast<int>(ply)))
 		return GameValues::Draw;
 
 	// Er vi naaet til bunden af traeet ?
